+++ sound/hda/codecs/ca0132.c
@@ -5557,7 +5557,7 @@ static int ca0132_vnode_switch_set(struct snd_kcontrol *kcontrol,
 	if (effective) {
 		int dir = get_amp_direction(kcontrol);
 		int ch = get_amp_channels(kcontrol);
-		unsigned long pval;
+		uintptr_t pval;
 
 		guard(mutex)(&codec->control_mutex);
 		pval = kcontrol->private_value;
@@ -6426,7 +6426,7 @@ static int ca0132_volume_info(struct snd_kcontrol *kcontrol,
 	hda_nid_t nid = get_amp_nid(kcontrol);
 	int ch = get_amp_channels(kcontrol);
 	int dir = get_amp_direction(kcontrol);
-	unsigned long pval;
+	uintptr_t pval;
 	int err;
 
 	switch (nid) {
@@ -6503,7 +6503,7 @@ static int ca0132_volume_put(struct snd_kcontrol *kcontrol,
 	effective = ca0132_is_vnode_effective(codec, nid, &shared_nid);
 	if (effective) {
 		int dir = get_amp_direction(kcontrol);
-		unsigned long pval;
+		uintptr_t pval;
 
 		CLASS(snd_hda_power, pm)(codec);
 		guard(mutex)(&codec->control_mutex);
@@ -6565,7 +6565,7 @@ static int ca0132_volume_tlv(struct snd_kcontrol *kcontrol, int op_flag,
 	hda_nid_t nid = get_amp_nid(kcontrol);
 	int ch = get_amp_channels(kcontrol);
 	int dir = get_amp_direction(kcontrol);
-	unsigned long pval;
+	uintptr_t pval;
 	int err;
 
 	switch (nid) {
